[W1 수정] DB 및 메서드 개선

+++ src/main/java/com/pet/comes/dto/Req/ScheduleDto.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 @Builder
 @NoArgsConstructor
 public class ScheduleDto {
-	private long iconId;
+	private int iconId;
 	private long scheduleId;
 	private long userId;
 	private String date;
